requirements.txt
markdown==3.7
Jinja2==3.1.6
python-frontmatter==1.1.0
